Refact : 컴포넌트를 분리& 폴더를 통일하여 유지보수성을 높임
